+++ src/Round.java
@@ -1,5 +1,5 @@
 public class Round {
-    public static Integer call(Double a) {
-        return (int) Math.round(a);
+    public static Double call(Double a) {
+        return (double) Math.round(a);
     }
 }
