+++ src/Subtraction.java
@@ -1,5 +1,5 @@
 public class Subtraction {
-    public static Integer call(Integer a, Integer b) {
+    public static Double call(Double a, Double b) {
         return a-b;
     }
 }
